changes completed and cards looks great
