+++ src/components/Home/Home.css
@@ -1,8 +1,8 @@
-.user-page {
+.home-page {
     display: flex;
-    grid-gap: 15px 15px;
-    justify-content:center;
-    padding: 20px;
+    grid-gap: 40px;
+    justify-content: center;
+    padding-top: 20px;
     padding-bottom: 100px;
 }
