+++ src/components/HomeProgramsDisplay/ListOfPrograms.css
@@ -1,0 +1,67 @@
+div#container.top-programs {
+    width: 1200px;
+    font-size: 22px;
+    /* box-shadow: 0 0 20px rgba(0, 0, 0, 0.15),0 0 20px rgba(0, 0, 0, 0.15),0 0 20px rgba(0, 0, 0, 0.15); */
+    /* backdrop-filter: blur(7px); */
+    height: fit-content;
+    border-radius: 5px;
+    /* border: 1px black solid; */
+}
+
+div#card-section{
+    display:grid;
+    grid-template-columns: 600px 600px;
+    column-gap: 20px;
+    row-gap: 20px;
+}
+
+
+img.cards-image{
+    border-radius: 5px 0px 0px 5px;
+    width: 300px;
+    height: 270px;
+    object-fit: fill;
+}
+
+.card {
+    display: flex;
+    background-color: #FFFFFF;
+    border-radius: 5px;
+    box-shadow: 0px 4px 4px rgba(2, 3, 4, 0.50);
+    height: fit-content;
+    cursor: pointer;
+    height: 270px;
+}
+
+.card div#img{
+}   
+
+div.card div.card-information{
+    padding-left: 10px;
+    background-color: #FFFFFF;
+    text-align: left;
+    width: 350px;
+    max-height: 270px;
+    border-radius: 0px 5px 5px 0px;
+}
+
+div.card div p {
+    font-size: 15px;
+}
+
+h4 {
+    margin-bottom: 12px;
+}
+
+h5 {
+    margin-bottom: 12px;
+}
+
+h6, .h4, .h5, .h6 {
+    margin-top: 0;
+    margin-bottom: 0;
+}
+
+
+
+
